fix vote-count tuple annotation and stale test comment
Fix round 1 from code review: the order-key tuple in
ClassifierLabelListByEntryView.post was still typed as int for the vote
count even though vote_counts is now dict[UUID, float], and its -1
sentinel didn't match. Also corrects a stale test comment that no longer
matched the code it described.

+++ api/views/classifier_label.py
@@ -115,11 +115,11 @@ class ClassifierLabelListByEntryView(APIView):
 
         for feed_entry_uuid, vote_counts in classifier_label_vote_counts.items():
             classifier_labels_with_order_keys: list[
-                tuple[ClassifierLabel, tuple[int, float]]
+                tuple[ClassifierLabel, tuple[float, float]]
             ] = [
                 (
                     classifier_label,
-                    (vote_counts.get(classifier_label.uuid, -1), random.random()),
+                    (vote_counts.get(classifier_label.uuid, -1.0), random.random()),
                 )
                 for classifier_label in classifier_labels
             ]
